requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
black==24.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
coverage==7.4.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
immutabledict==4.2.0
importlab==0.8.1
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.5.0
jupyter_core==5.5.0
libcst==1.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
msgspec==0.18.6
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
pluggy==1.4.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycnite==2023.10.11
pydot==2.0.0
Pygments==2.16.1
pyparsing==3.1.2
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytype==2024.4.11
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-mock==1.12.1
ruff==0.3.7
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
stack-data==0.6.3
tabulate==0.9.0
taskipy==1.12.2
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.2.1
wcwidth==0.2.9
xmltodict==0.13.0
zipp==3.17.0
